subida clase 08/10

+++ Ejercicios/src/tema3_bucles/Ejercicio01_bloque2_Media.java
@@ -7,9 +7,10 @@ public class Ejercicio01_bloque2_Media {
 	public static void main(String[] args) {
 		//pedimos al usuario la cantidad de números sobre los que quiere realizar la media
 		int mediaNumeros = Integer.parseInt(JOptionPane.showInputDialog("Introduzca cantidad de números de los que quiere realizar la media"));
-		int suma = 0;
+		float suma = 0;
 		//realizamos bucle for
 		for (int i=0; i < mediaNumeros; i++) {
+			//pedimos al usuario que introduzca los números que vamos a sumar para posteriormente realizar su media
 			int numIntroducidos = Integer.parseInt(JOptionPane.showInputDialog("Introduzca números para realizar media"));
 			suma +=numIntroducidos;
